requirements.txt
appdirs==1.4.3
asgiref==3.2.7
autopep8==1.5.2
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.6
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-js-asset==1.2.2
django-multiselectfield==0.1.12
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
pep8==1.7.1
progress==1.5
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
webencodings==0.5.1
